Add files via upload
This is the wiring diagram of the relay circuits. Couple of things to note here is that we are using an active low system so the relays will be in on position when we give a low in code. This is important to note because the same logic has been used in our code for the microcontroller. Next main thing to note here is that we are directly plugging the relay into our existing switches this way we can still use the physical switches and also we can control them digitally by just overriding the physical switch.